stubbedout navbar navigation menu in fragments

+++ src/main/java/com/nendejan/lc101FinalProject/controllers/ShiftController.java
@@ -5,7 +5,7 @@ import com.nendejan.lc101FinalProject.models.Schedule;
 import com.nendejan.lc101FinalProject.models.Shift;
 import com.nendejan.lc101FinalProject.models.User;
 import com.nendejan.lc101FinalProject.models.data.*;
-import com.nendejan.lc101FinalProject.models.forms.AddShiftEmployeeForm;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -49,8 +49,16 @@ public class ShiftController {
         int loggedInUserId = Integer.parseInt(cookieValue);
 
         User loggedInUser = userDao.findOne(loggedInUserId);
+        ArrayList<Shift> workPlacePrimaryShifts =  new ArrayList<>();
 
-        model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplaceShifts());
+        for(Shift shift : loggedInUser.getWorkplace().getWorkplaceShifts()){
+            if(shift.isPrimaryShift() == true){
+                workPlacePrimaryShifts.add(shift);
+            }
+
+        }
+
+        model.addAttribute("shifts", workPlacePrimaryShifts);
         model.addAttribute("title", "Shifts for Schedule");
 
 
@@ -80,7 +88,7 @@ public class ShiftController {
             model.addAttribute("title", "Add Shift");
             return "shifts/add";
         }
-
+        shift.setPrimaryShift(true);
         loggedInUser.getWorkplace().addShift(shift);
         shiftDao.save(shift);
         return "redirect:";
@@ -101,40 +109,59 @@ public class ShiftController {
             model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplaceShifts());
             model.addAttribute("title", "Remove Shifts");
 
+
             return "shifts/index";
         }
 
 
+
+
         model.addAttribute("hasShifts", true);
-        model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplaceShifts());
+        model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplacePrimaryShifts(loggedInUser.getWorkplace()));
         model.addAttribute("title", "Remove Shifts");
         return "shifts/remove";
 
     }
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
-    public String processRemoveShiftForm(Model model, HttpServletRequest request, @CookieValue(value="loggedInCookie") String cookieValue, @RequestParam List<Integer> shiftsToRemove) {
+    public String processRemoveShiftForm(Model model, HttpServletRequest request, @CookieValue(value="loggedInCookie") String cookieValue, @RequestParam(required = false) List<Integer> shiftsToRemove) {
 
         int loggedInUserId = Integer.parseInt(cookieValue);
 
         User loggedInUser = userDao.findOne(loggedInUserId);
 
-        model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplaceShifts());
+        model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplacePrimaryShifts(loggedInUser.getWorkplace()));
         model.addAttribute("title", "Remove Shift");
 
+
+        if(shiftsToRemove==null){
+            model.addAttribute("hasShifts", true);
+            model.addAttribute("shifts", loggedInUser.getWorkplace().getWorkplacePrimaryShifts(loggedInUser.getWorkplace()));
+            model.addAttribute("title", "Remove Shift");
+
+            model.addAttribute("noShiftSelected", "You must select a shift to remove.");
+
+            return "shifts/remove";
+
+        }
+        if(shiftsToRemove !=null){
         for (Integer removeMeId : shiftsToRemove){
             Shift removeMe = shiftDao.findOne(removeMeId);
             loggedInUser.getWorkplace().removeShift(removeMe);
 
-            for(Schedule schedule :loggedInUser.getWorkplace().getWorkplaceSchedules()){
+            for(Schedule schedule : loggedInUser.getWorkplace().getWorkplaceSchedules()){
                 schedule.removeShift(removeMe);
+            }
+            for(Employee employee : loggedInUser.getWorkplace().getEmployeeRoster()){
+                employee.removeShiftFromScheduledShifts(removeMe);
+                employee.removeShiftAvailability(removeMe);
             }
 
 
             shiftDao.delete(removeMeId);
+            return "redirect:/login/shifts/remove";
 
-
-        }
+        }}
         return "redirect:/login/shifts/remove";
     }
